Move `DendriteSegementsBooster` to frameworks.
