+++ nupic/research/frameworks/dendrites/__init__.py
@@ -22,3 +22,4 @@
 from .modules import *
 from .metrics import *
 from .apply_dendrites_hook import *
+from .dendrite_segments_booster import *
